+++ src/format/sd_jwt/store.rs
@@ -7,7 +7,7 @@ use serde_json::Value;
 use crate::core::Kind;
 use crate::format::FormatProfile;
 use crate::format::sd_jwt::{Disclosure, verify};
-use crate::oid4vp::types::{Claim, Queryable};
+use crate::oid4vp::verifier::{Claim, Queryable};
 
 /// Convert a `dc+sd-jwt` encoded credential to a `Queryable` object.
 ///
