+++ src/oid4vp/handlers/create_request.rs
@@ -8,11 +8,11 @@ use anyhow::Context;
 use chrono::Utc;
 
 use crate::generate;
-use crate::oid4vp::endpoint::{Body, Error, Handler, Request, Response, Result};
+use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::error::invalid;
 use crate::oid4vp::provider::{Metadata, Provider, StateStore};
 use crate::oid4vp::state::Expire;
-use crate::oid4vp::types::{
+use crate::oid4vp::verifier::{
     ClientId, DeviceFlow, GenerateRequest, GenerateResponse, RequestObject, ResponseType,
 };
 use crate::state::State;
