+++ src/oid4vp/handlers/response.rs
@@ -25,10 +25,10 @@ use anyhow::Context;
 
 use crate::core::Kind;
 use crate::format::{mso_mdoc, sd_jwt, w3c_vc};
-use crate::oid4vp::endpoint::{Body, Error, Handler, Request, Response, Result};
+use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::error::invalid;
 use crate::oid4vp::provider::{Provider, StateStore};
-use crate::oid4vp::types::{
+use crate::oid4vp::verifier::{
     AuthorzationResponse, Queryable, RedirectResponse, RequestObject, RequestedFormat,
 };
