+++ src/oid4vp/verifier/request.rs
@@ -13,7 +13,7 @@ use serde_json::Value;
 use crate::Kind;
 use crate::core::urlencode;
 use crate::oid4vp::JwtType;
-use crate::oid4vp::types::{DcqlQuery, VerifierMetadata, Wallet};
+use crate::oid4vp::verifier::{DcqlQuery, VerifierMetadata, Wallet};
 
 const UNRESERVED: &AsciiSet =
     &NON_ALPHANUMERIC.remove(b'&').remove(b'=').remove(b'.').remove(b'_').remove(b'-').remove(b'~');
